Add office config API and client endpoints also added the office config edit options in admin dashboard)
Introduce office configuration management: add OfficeConfigController and DTOs to expose GET and PUT /api/office-config with validation, time parsing, mapping, and logging. Add UpsertActiveAsync to CompanyPolicyRepository and its interface to create or update the active policy. Adjust AttendanceService hours accumulation to cast TotalHours to int (accumulate whole hours). Add corresponding client API functions (getOfficeConfig, updateOfficeConfig) and tidy comments in api.js.

+++ SPL.Attendance.Business/Services/AttendanceService.cs
@@ -325,7 +325,7 @@ namespace SPL.Attendance.Business.Services
             {
                 if (log.CheckInTime.HasValue && log.CheckOutTime.HasValue)
                 {
-                    totalHours += (decimal)(log.CheckOutTime.Value
+                    totalHours += (int)(log.CheckOutTime.Value
                                  - log.CheckInTime.Value).TotalHours;
                 }
             }
